+++ assetconnection/mqtt/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/mqtt/MqttAssetConnectionConfigTest.java
@@ -22,7 +22,6 @@ public class MqttAssetConnectionConfigTest {
 
     @Test
     public void testEquals() {
-        EqualsVerifier.simple().forClass(MqttAssetConnectionConfig.class)
-                .verify();
+        EqualsVerifier.simple().forClass(MqttAssetConnectionConfig.class).verify();
     }
 }
