+++ assetconnection/mqtt/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/mqtt/provider/config/MqttSubscriptionProviderConfigTest.java
@@ -22,7 +22,6 @@ public class MqttSubscriptionProviderConfigTest {
 
     @Test
     public void testEquals() {
-        EqualsVerifier.simple().forClass(MqttSubscriptionProviderConfig.class)
-                .verify();
+        EqualsVerifier.simple().forClass(MqttSubscriptionProviderConfig.class).verify();
     }
 }
