+++ assetconnection/opcua/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/opcua/provider/config/OpcUaOperationProviderConfigTest.java
@@ -22,7 +22,6 @@ public class OpcUaOperationProviderConfigTest {
 
     @Test
     public void testEquals() {
-        EqualsVerifier.simple().forClass(OpcUaOperationProviderConfig.class)
-                .verify();
+        EqualsVerifier.simple().forClass(OpcUaOperationProviderConfig.class).verify();
     }
 }
