+++ assetconnection/opcua/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/opcua/provider/config/OpcUaValueProviderConfigTest.java
@@ -22,7 +22,6 @@ public class OpcUaValueProviderConfigTest {
 
     @Test
     public void testEquals() {
-        EqualsVerifier.simple().forClass(OpcUaValueProviderConfig.class)
-                .verify();
+        EqualsVerifier.simple().forClass(OpcUaValueProviderConfig.class).verify();
     }
 }
